refactor(bff): move view contracts into application layer
