+++ packages/bff/src/application/index.ts
@@ -3,3 +3,4 @@ export * from "./orchestrator/mutation-orchestrator.service";
 export * from "./orchestrator/query-orchestrator.service";
 export * from "./orchestrator/query-pipeline.service";
 export * from "./view/temporary-view-adapter.service";
+export * from "./view/view.contract";
