+++ packages/bff/src/application/view/temporary-view-adapter.service.ts
@@ -15,7 +15,7 @@ import {
 import { MetaRegistryService } from "../../interface/gateway/meta-registry.service";
 import { OrgScopeService } from "../../infra/integration/org-scope.service";
 import { PostgresQueryExecutorService } from "../../infra/integration/postgres-query-executor.service";
-import type { ViewApiRequest } from "../../types";
+import type { ViewApiRequest } from "./view.contract";
 
 export interface TemporaryViewExecutionResult {
   requestId: string;
