+++ packages/bff/src/application/view/view.contract.ts
